Add files via upload
i ran the last version of the file through online software to make the background transparent and then converted it from a png to a svg file. The color of the bee cartoon was unfortunately stripped during processing--if you'd like, I can add it back in there. Let's discuss this further in a bit.